pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "lojaflow"
version = "1.0.0"
description = "Sistema de gestão para pequenos comércios"
requires-python = ">=3.11"
dependencies = [
    "PySide6>=6.6.0",
    "SQLAlchemy>=2.0.0",
    "python-escpos>=3.0.0",
    "matplotlib>=3.8.0",
    "Pillow>=10.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.pyinstaller]
name = "LojaFlow"
onefile = true
windowed = true
icon = "assets/logo.png"
